+++ client.go
@@ -172,14 +172,15 @@ func New(config Config, opts ...Option) (*Client, error) {
 		}
 	}
 	client.blob, err = blob.New(blob.Config{
-		Master:           client.master,
-		HTTPClient:       applied.httpClient,
-		UserAgent:        config.UserAgent,
-		BearerToken:      config.BearerToken,
-		Retry:            config.Retry,
-		EndpointPolicy:   endpointPolicyOrDefault(config.BlobEndpointPolicy, config.EndpointPolicy),
-		UsePublicURLs:    config.UsePublicURLs,
-		LocationCacheTTL: config.BlobLocationCacheTTL,
+		Master:                    client.master,
+		HTTPClient:                applied.httpClient,
+		UserAgent:                 config.UserAgent,
+		BearerToken:               config.BearerToken,
+		Retry:                     config.Retry,
+		EndpointPolicy:            endpointPolicyOrDefault(config.BlobEndpointPolicy, config.EndpointPolicy),
+		UsePublicURLs:             config.UsePublicURLs,
+		LocationCacheTTL:          config.BlobLocationCacheTTL,
+		EnableVolumeAuthorization: config.EnableVolumeAuthorization,
 	})
 	if err != nil {
 		return nil, err